LessonRest

+++ Daw/src/main/java/com/urjc/daw/models/lessons/Lesson.java
@@ -41,7 +41,7 @@ public class Lesson {
         return id;
     }
 
-    public void setIdLesson(Integer idLesson) {
+    public void setIdLesson(long idLesson) {
         this.id = idLesson;
     }
